Reduce duplication in str/eq implementations between struct and union.
